backend/requirements.txt
# DocuMind Backend Dependencies
# FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.1.0

# Vector database and embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# AI and ML
torch>=2.1.1
numpy>=1.24.0
transformers>=4.35.2

# Database
motor>=3.3.2
pymongo>=4.6.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT>=2.8.0
